% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Road accident map by state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer or character}

\item{year}{Integer or character}
}
\value{
Prints invalid state number if state number does not exist or 'no accidents'
if accidents don't exist, and a map if both statements are valid and all other
parameters exist.
}
\description{
This function prints a map of road accidents for a state and year(s) given by users.
}
\examples{
\dontrun{
fars_map_state(1,"2014")
fars_map_state(1,2014)
}
}
