% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in data.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path and name of data set to download.}
}
\value{
This function either prints an error message or returns a data frame object depending on whether the data set is
found (and can be downloaded).
}
\description{
This function is given a filename that has both path and name of dataset to download using read_csv function.
The function first checks if the data exist, provides an error message if they do not, and downloads the data
if they do. The data object is in the form of data frame. Note that dplyr::tbl_df should be tibble::as_tibble
as dplyr::tbl_df has become obsolete. This changes import from dplyr to tibble for this function.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
