% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Download multiple files.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years given by user that specifies files to download.}
}
\value{
A list of data frames containing data of specified years. In case of invalid year a warning and a NULL.
}
\description{
This function will download several years of data on one go from a list of years given by user.
In case of any error, the function prints warnings of invalid years and returns NULL.
}
\examples{
\dontrun{
fars_read_years(c("2013","2014","2015"))
fars_read_years(c(2013,2014,2015))
}
}
