% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes multiple years of data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years to download and summarize.}
}
\value{
a summary in data frame format of n by years and months.
}
\description{
This function reads given years of data and returns a summary in data frame format of n by years and months.
}
\examples{
\dontrun{
fars_summarize_years(c("2013","2014","2015"))
fars_summarize_years(c(2013,2014,2015))
}
}
