% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or an integer provided by user that specifies the year of the file name.}
}
\value{
A character string of a pre set name and user given year.
}
\description{
This function returns a filename in C-style string formatting from a character or integer year given by the user and a pre set file name.
If the year is provided as a character string, the function will convert it to integer before returning the filename.
}
\examples{
make_filename(2014)
make_filename("2014")
}
